Adding the Program.cs file version 1.0
In this commit is added the Program.cs file which is the start one. It configures the connection with the SQL Server by using the AddDbContext Service provided by ASP.Net Core. This is the first version of the file (version 1.0) and lots of functionality will be added to it!